# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RAKT KAVACH - National Digital Blood Grid API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hospitals:
    get:
      operationId: listHospitals
      tags: [hospitals]
      summary: List hospitals
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Hospitals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HospitalListResponse"
    post:
      operationId: createHospital
      tags: [hospitals]
      summary: Create hospital
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HospitalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hospital"
  /hospitals/{id}:
    get:
      operationId: getHospital
      tags: [hospitals]
      summary: Get hospital
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hospital
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hospital"
  /hospitals/{id}/inventory:
    get:
      operationId: getHospitalInventory
      tags: [hospitals]
      summary: Get hospital blood inventory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryListResponse"
  /hospitals/{id}/requests:
    get:
      operationId: getHospitalRequests
      tags: [hospitals]
      summary: Get hospital blood requests
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequestListResponse"

  # BLOOD BANKS
  /blood-banks:
    get:
      operationId: listBloodBanks
      tags: [blood-banks]
      summary: List blood banks
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: bloodGroup
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Blood banks list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodBankListResponse"
    post:
      operationId: createBloodBank
      tags: [blood-banks]
      summary: Create blood bank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodBankInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodBank"
  /blood-banks/{id}:
    get:
      operationId: getBloodBank
      tags: [blood-banks]
      summary: Get blood bank
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blood bank
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodBank"
  /blood-banks/{id}/inventory:
    get:
      operationId: getBloodBankInventory
      tags: [blood-banks]
      summary: Get blood bank inventory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryListResponse"
    post:
      operationId: addBloodBankInventory
      tags: [blood-banks]
      summary: Add inventory to blood bank
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inventory"

  # LABORATORIES
  /laboratories:
    get:
      operationId: listLaboratories
      tags: [laboratories]
      summary: List laboratories
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Laboratories list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaboratoryListResponse"
    post:
      operationId: createLaboratory
      tags: [laboratories]
      summary: Create laboratory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LaboratoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Laboratory"
  /laboratories/{id}:
    get:
      operationId: getLaboratory
      tags: [laboratories]
      summary: Get laboratory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Laboratory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Laboratory"
  /laboratories/{id}/tests:
    get:
      operationId: getLaboratoryTests
      tags: [laboratories]
      summary: Get laboratory test results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabTestListResponse"
    post:
      operationId: createLabTest
      tags: [laboratories]
      summary: Create lab test
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabTestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabTest"

  # AMBULANCES
  /ambulances:
    get:
      operationId: listAmbulances
      tags: [ambulances]
      summary: List ambulances
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Ambulances list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AmbulanceListResponse"
    post:
      operationId: createAmbulance
      tags: [ambulances]
      summary: Create ambulance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmbulanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ambulance"
  /ambulances/{id}:
    get:
      operationId: getAmbulance
      tags: [ambulances]
      summary: Get ambulance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ambulance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ambulance"
    patch:
      operationId: updateAmbulance
      tags: [ambulances]
      summary: Update ambulance status/location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmbulanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ambulance"

  # VOLUNTEERS
  /volunteers:
    get:
      operationId: listVolunteers
      tags: [volunteers]
      summary: List volunteers
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Volunteers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerListResponse"
    post:
      operationId: createVolunteer
      tags: [volunteers]
      summary: Register as volunteer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
  /volunteers/{id}:
    get:
      operationId: getVolunteer
      tags: [volunteers]
      summary: Get volunteer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Volunteer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"

  # BLOOD INVENTORY
...
components:
  schemas:
...
    HospitalInput:
      type: object
      required: [name, registrationNumber, state, district, address, contactNumber, type]
      properties:
        name:
          type: string
        registrationNumber:
          type: string
        state:
          type: string
        district:
          type: string
        address:
          type: string
        contactNumber:
          type: string
        email:
          type: string
        type:
          type: string
          enum: [GOVERNMENT, PRIVATE, TRUST, MILITARY]
        totalBeds:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
    Hospital:
      type: object
      required: [id, name, registrationNumber, state, district, address, contactNumber, type, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        registrationNumber:
          type: string
        state:
          type: string
        district:
          type: string
        address:
          type: string
        contactNumber:
          type: string
        email:
          type: ["string", "null"]
        type:
          type: string
        totalBeds:
          type: ["integer", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    HospitalListResponse:
      type: object
      required: [hospitals, total, page, limit]
      properties:
        hospitals:
          type: array
          items:
            $ref: "#/components/schemas/Hospital"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # BLOOD BANK SCHEMAS
    BloodBankInput:
      type: object
      required: [name, licenseNumber, state, district, address, contactNumber]
      properties:
        name:
          type: string
        licenseNumber:
          type: string
        state:
          type: string
        district:
          type: string
        address:
          type: string
        contactNumber:
          type: string
        email:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        isComponentFacility:
          type: boolean
        is24x7:
          type: boolean
    BloodBank:
      type: object
      required: [id, name, licenseNumber, state, district, address, contactNumber, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        licenseNumber:
          type: string
        state:
          type: string
        district:
          type: string
        address:
          type: string
        contactNumber:
          type: string
        email:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        isComponentFacility:
          type: boolean
        is24x7:
          type: boolean
        isActive:
          type: boolean
        totalUnitsInStock:
          type: integer
        createdAt:
          type: string
    BloodBankListResponse:
      type: object
      required: [bloodBanks, total, page, limit]
      properties:
        bloodBanks:
          type: array
          items:
            $ref: "#/components/schemas/BloodBank"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # LABORATORY SCHEMAS
    LaboratoryInput:
      type: object
      required: [name, licenseNumber, state, district, address, contactNumber]
      properties:
        name:
          type: string
        licenseNumber:
          type: string
        state:
          type: string
        district:
          type: string
        address:
          type: string
        contactNumber:
          type: string
        email:
          type: string
    Laboratory:
      type: object
      required: [id, name, licenseNumber, state, district, address, contactNumber, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        licenseNumber:
          type: string
        state:
          type: string
        district:
          type: string
        address:
          type: string
        contactNumber:
          type: string
        email:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    LaboratoryListResponse:
      type: object
      required: [laboratories, total, page, limit]
      properties:
        laboratories:
          type: array
          items:
            $ref: "#/components/schemas/Laboratory"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    LabTestInput:
      type: object
      required: [bloodUnitId, testType, result]
      properties:
        bloodUnitId:
          type: integer
        testType:
          type: string
          enum: [HIV, HBV, HCV, SYPHILIS, MALARIA, BLOOD_GROUP_CONFIRM, CROSSMATCH]
        result:
          type: string
          enum: [POSITIVE, NEGATIVE, INCONCLUSIVE]
        notes:
          type: string
    LabTest:
      type: object
      required: [id, laboratoryId, bloodUnitId, testType, result, testedAt]
      properties:
        id:
          type: integer
        laboratoryId:
          type: integer
        bloodUnitId:
          type: integer
        testType:
          type: string
        result:
          type: string
        notes:
          type: ["string", "null"]
        testedAt:
          type: string
        testedBy:
          type: ["string", "null"]
    LabTestListResponse:
      type: object
      required: [tests, total]
      properties:
        tests:
          type: array
          items:
            $ref: "#/components/schemas/LabTest"
        total:
          type: integer

    # AMBULANCE SCHEMAS
    AmbulanceInput:
      type: object
      required: [vehicleNumber, state, district, contactNumber, type]
      properties:
        vehicleNumber:
          type: string
        state:
          type: string
        district:
          type: string
        contactNumber:
          type: string
        type:
          type: string
          enum: [BASIC, ADVANCED, MOBILE_ICU, BLOOD_TRANSPORT]
        driverName:
          type: string
        latitude:
          type: number
        longitude:
          type: number
...
    Ambulance:
      type: object
      required: [id, vehicleNumber, state, district, contactNumber, type, status, createdAt]
      properties:
        id:
          type: integer
        vehicleNumber:
          type: string
        state:
          type: string
        district:
          type: string
        contactNumber:
          type: string
        type:
          type: string
        driverName:
          type: ["string", "null"]
        status:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        emergencyId:
          type: ["integer", "null"]
        createdAt:
          type: string
    AmbulanceListResponse:
      type: object
      required: [ambulances, total, page, limit]
      properties:
        ambulances:
          type: array
          items:
            $ref: "#/components/schemas/Ambulance"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # VOLUNTEER SCHEMAS
...
    InventoryInput:
      type: object
      required: [bloodGroup, units, collectionDate]
      properties:
        bloodGroup:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        units:
          type: number
        collectionDate:
          type: string
        expiryDate:
          type: string
        donorId:
          type: integer
    Inventory:
      type: object
      required: [id, bloodBankId, bloodGroup, units, collectionDate, expiryDate, status, createdAt]
      properties:
        id:
          type: integer
        bloodBankId:
          type: integer
        bloodBankName:
          type: string
        bloodGroup:
          type: string
        units:
          type: number
        collectionDate:
          type: string
        expiryDate:
          type: string
        status:
          type: string
          enum: [AVAILABLE, RESERVED, DISPATCHED, EXPIRED, DISCARDED]
        donorId:
          type: ["integer", "null"]
        createdAt:
          type: string
    InventoryListResponse:
      type: object
      required: [inventory, total, page, limit]
      properties:
        inventory:
          type: array
          items:
            $ref: "#/components/schemas/Inventory"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    BloodRequestListResponse:
      type: object
      required: [requests, total, page, limit]
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/BloodRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # EMERGENCY SCHEMAS
...